More refactoring, cleaning up color, optimizing FPS

+++ src/sim/VertexTexture.h
@@ -3,10 +3,10 @@
 namespace mms {
 
 struct VertexTexture {
-    double x; // x position
-    double y; // y position
-    double u; // u position (x position in the texture)
-    double v; // v position (y position in the texture)
+    float x; // x position
+    float y; // y position
+    float u; // u position (x position in the texture)
+    float v; // v position (y position in the texture)
 };
 
 }
